examples/spark-scala: removed float casts to logMetric

+++ examples/spark-scala/src/main/scala/org/andre/mlflow/examples/TrainDecisionTree.scala
@@ -106,7 +106,7 @@ object TrainDecisionTree {
     println(s"Learned regression tree model:\n ${treeModel.toDebugString}")
 
     // MLflow - Log metric
-    mlflowClient.logMetric(runId, "rmse",rmse.toFloat)
+    mlflowClient.logMetric(runId, "rmse",rmse)
 
     // MLflow - Log simple artifact
     val path="info.txt"
